got the proficiency component working and styled
