+++ client/components/Proficiency.scss
@@ -1,0 +1,32 @@
+@import '../variables';
+
+.proficiency {
+  * {
+    color: $color5;
+  }
+  background-color: $color4;
+  padding: 0px;
+  width: 115px;
+  height: 115px;
+  border: 2px solid $color6;
+  border-radius: 10%;
+  overflow: hidden;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+
+  .proficiency-title {
+    padding: 2px;
+    width: 100%;
+    text-align: center;
+    color: $color4;
+    background-color: $color5;
+  }
+  .proficiency-total {
+    height: 100%;
+    text-align: center;
+    input {
+      margin: 33% 0 33% 0;
+    }
+  }
+}
